Drop THIS_MODULE from class_create
As-is, this FTBFS on 6.8.12-amd64. There must have been a change
recently, I checked for similar class_create callsites and this seemed
to have worked for me. I'm not tracking these changes well enough to
know if this is a good idea or not, but figured i'd offer the patch up.